Refactor outxml_parser to use the FleurXMLContext
The ParseTasks helper class is moved into the same module and made private
(Also slimmed down by removing the functionality to completely replace everything)

The decorator register_parsing_function is removed and replaced by a simple
mapping to the methods of the context (This decorator was one of the main reasons
for cyclic import problems cropping up)

During refactoring a bug was found not correctly choosing the last stable iteration
for broken XML files when there are more than one iteration
