+++ .pre-commit-config.yaml
@@ -92,11 +92,9 @@ repos:
             masci_tools/util/lockable_containers.py|
             masci_tools/util/case_insensitive_dict.py|
             masci_tools/util/schema_dict_util.py|
-            masci_tools/util/parse_task_decorators.py|
             masci_tools/util/logging_util.py|
             masci_tools/util/python_util.py|
             masci_tools/util/math_util.py|
-            masci_tools/util/parse_tasks.py|
             masci_tools/util/parse_utils.py|
             masci_tools/util/typing.py|
             masci_tools/util/econfig.py|
