+++ masci_tools/io/parsers/fleur/task_migrations.py
@@ -15,8 +15,8 @@ In this module migration functions for the task definitions are collected
 from __future__ import annotations
 
 import copy
-from masci_tools.util.parse_tasks_decorators import register_migration
 from typing import Any
+from . import register_migration
 
 
 @register_migration(base_version='0.33', target_version=['0.31', '0.30', '0.29'])
